backend/requirements.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7

# API and HTTP
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.1.1
numpy==1.24.3
ta-lib==0.4.28
yfinance==0.2.22

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6

# Real-time and Background Tasks
celery==5.3.2
redis==4.6.0
eventlet==0.33.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# Development
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
